+++ packages/codegen/src/get-used-types.ts
@@ -85,17 +85,17 @@ export const getUsedTypes = (
     })
     buildEnum(
       "args",
-      pallet.calls,
+      pallet.calls?.type,
       (name) => builder.buildCall(pallet.name, name)!,
     )
     buildEnum(
       "values",
-      pallet.events,
+      pallet.events?.type,
       (name) => builder.buildEvent(pallet.name, name)!,
     )
     buildEnum(
       "values",
-      pallet.errors,
+      pallet.errors?.type,
       (name) => builder.buildError(pallet.name, name)!,
     )
   })