+++ packages/metadata-builders/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Unreleased
 
+### Changed
+
+- `getLookupFn` takes a `UnifiedMetadata`, instead of `V14 | V15`.
+
 ## 0.11.0 - 2025-04-24
 
 ### Added